Rewrite the validate
The sorting from the JS won't rewrite the id's of the array. This mean we have to reorder the id's of the post. We got the post in the right order, but the id's of the array aren't from 1 to 9. So we rewrite them and reorder the errors of the widget.