Stop endless loop in generate with new wizard
When creating a new wizard Column, the executePostAction hook was called again and again when used with the DcG. To stop this loop, the $_SERVER['HTTP_X_REQUESTED_WITH'] will be set to null before the DcG call. So the executePostAction hook will not be called again in the DcG. After DcG creation, the value is set back to its original value.